Remove integration tests targets
