+++ Tests/TuistCacheTests/CacheGraphContentHasherIntegrationTests.swift
