+++ Tests/TuistCoreTests/MetadataProviders/FrameworkMetadataProviderIntegrationTests.swift
