+++ Tests/TuistCoreTests/MetadataProviders/PrecompiledMetadataProviderIntegrationTests.swift
