+++ Tests/TuistGeneratorTests/Generator/MultipleConfigurationsIntegrationTests.swift
