+++ Tests/TuistGeneratorTests/Generator/StableStructureIntegrationTests.swift
