+++ Tests/TuistGeneratorTests/Generator/SwiftPackageManagerInteractorTests.swift
