+++ Tests/TuistGeneratorTests/Generator/WorkspaceGeneratorIntegrationTests.swift
