+++ Tests/TuistGeneratorTests/Generator/XcodeProjWriterTests.swift
