+++ Tests/TuistKitTests/Commands/DumpServiceIntegrationTests.swift
