+++ Tests/TuistMigrationTests/Utilities/EmptyBuildSettingsCheckerIntegrationTests.swift
