+++ Tests/TuistMigrationTests/Utilities/SettingsToXCConfigExtractorIntegrationTests.swift
