+++ Tests/TuistMigrationTests/Utilities/TargetsExtractorIntegrationTests.swift
