+++ Tests/TuistScaffoldTests/Utils/TemplatesDirectoryLocatorIntegrationTests.swift
